pyproject.toml
[tool.poetry]
name = "wise-api"
version = "0.1.0"
description = "API Services for Wise application"
authors = ["muhafiq <muhafiq18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-restful = "^0.3.10"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
flask-marshmallow = "^1.2.1"
marshmallow-sqlalchemy = "^1.1.0"
flask-migrate = "^4.0.7"
pymysql = "^1.1.1"
flask-bcrypt = "^1.0.1"
flask-jwt-extended = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
